Create README.md
I've just started learning Blazor and as I'm learning I like to collect useful info from around the web and give what I find a go. Sometimes the examples I find can be a bit daunting and sometimes they are concept based, don't quite work if pasted in verbatim and need a little extra work. So to this end, I like to make my own simple examples based on what I've found. Hopefully these examples might be of use to others. They are all super basic and may not be 100% correct (i'm still learning), but they do exactly what they are supposed to do without any frills. I should say again, these are my notes to myself, for when I forget how to do certain tasks...

Take a look at the code in this site and it should help you on your way

Here are 3 examples of passing data from page to page (or page to component). There are others ways, such as passing using routing with lots of simple examples available on the Web and still other ways which can get quite complex. I like to start with as simple as possible then add my own stuff to them once I understand better what I'm doing.

Example 1 is shown within this page. A class with data is passed from index.razor to and embedded page/component, ChildComponent.razor. It shows simple passing to components as parameters.

Example 2 is shown by clicking the counter page. The data is set in this index page and held in a scoped class which is injected into the counter page

Example 3 Not all functionality is available within Blazor. A good example would be Request.HttpContext.Connection.RemoteIpAddress, on a live Blazor server it doesn't work. Luckily it does work in Razor pages (cshtml). A work around is to use the _host.cshtm page to capture the data, and then pass the captured data to the child pages using Cascading values. You have to be a little careful as the _host.cshtm page isn't updated during blazor page changes.